refactor: calculator

+++ calculator.py
@@ -12,7 +12,7 @@ def calculator_option(operation, number_1,number_2):
         return (number_1 * number_2)
     elif operation == '/':
         print('{} / {} = '.format(number_1, number_2))
-        return (number_1 / number_2)
+        return int(number_1 / number_2)
     else:
         return ('You have not typed a valid operator')
